+++ packages/admin/admin-react-select/src/ReactSelect.tsx
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
+<<<<<<< HEAD
 import { createComponentSlot } from "@comet/admin";
 import { SvgIconComponent } from "@mui/icons-material";
 import CancelIcon from "@mui/icons-material/Cancel";
@@ -6,6 +7,12 @@ import ClearIcon from "@mui/icons-material/Clear";
 import DropdownIcon from "@mui/icons-material/KeyboardArrowDown";
 import { Chip, ComponentsOverrides, InputBase, inputBaseClasses, MenuItem, Paper, Theme, Typography, useTheme } from "@mui/material";
 import { css } from "@mui/material/styles";
+=======
+import { ChevronDown, Clear, Close } from "@comet/admin-icons";
+import { Chip, ComponentsOverrides, InputBase, InputBaseProps, MenuItem, Paper, SvgIconProps, Theme, Typography, useTheme } from "@mui/material";
+import { WithStyles, withStyles } from "@mui/styles";
+import classNames from "classnames";
+>>>>>>> main
 import * as React from "react";
 import Select, { OptionTypeBase } from "react-select";
 import AsyncSelect, { Props as ReactSelectAsyncProps } from "react-select/async";
@@ -185,7 +192,7 @@ function MultiValue<OptionType extends OptionTypeBase>(props: MultiValueProps<Op
             ownerState={{ focused: props.isFocused }}
             label={props.children}
             onDelete={props.removeProps.onClick}
-            deleteIcon={<CancelIcon {...props.removeProps} />}
+            deleteIcon={<Clear {...props.removeProps} />}
         />
     );
 }
@@ -247,7 +254,7 @@ function ClearIndicator<OptionType extends OptionTypeBase, IsMulti extends boole
     selectProps,
     clearValue,
 }: IndicatorProps<OptionType, IsMulti>) {
-    const Icon = selectProps.clearIcon ? selectProps.clearIcon : ClearIcon;
+    const Icon = selectProps.clearIcon ? selectProps.clearIcon : Close;
     return (
         <ClearIndicatorSlot onClick={clearValue}>
             <Icon fontSize="inherit" color="inherit" />
@@ -274,8 +281,8 @@ const DropdownIndicatorSlot = createComponentSlot("div")<SelectClassKey>({
 );
 
 function DropdownIndicator<OptionType extends OptionTypeBase, IsMulti extends boolean>({ selectProps }: IndicatorProps<OptionType, IsMulti>) {
-    const DefaultIcon = selectProps.dropdownIcon ? selectProps.dropdownIcon : DropdownIcon;
-    const OpenIcon = selectProps.dropdownIconOpen ? selectProps.dropdownIconOpen : DropdownIcon;
+    const DefaultIcon = selectProps.dropdownIcon ? selectProps.dropdownIcon : ChevronDown;
+    const OpenIcon = selectProps.dropdownIconOpen ? selectProps.dropdownIconOpen : ChevronDown;
     const Icon = selectProps.menuIsOpen ? OpenIcon : DefaultIcon;
 
     return (
@@ -300,11 +307,20 @@ const components = {
     DropdownIndicator,
 };
 
+<<<<<<< HEAD
 export interface SelectProps<OptionType extends OptionTypeBase, IsMulti extends boolean = false> {
     selectComponent: React.ComponentType<ReactSelectProps<OptionType, IsMulti>>;
     clearIcon?: SvgIconComponent;
     dropdownIcon?: SvgIconComponent;
     dropdownIconOpen?: SvgIconComponent;
+=======
+export interface SelectProps<OptionType extends OptionTypeBase> {
+    theme: Theme;
+    selectComponent: React.ComponentType<ReactSelectProps<OptionType>>;
+    clearIcon?: React.ComponentType<SvgIconProps>;
+    dropdownIcon?: React.ComponentType<SvgIconProps>;
+    dropdownIconOpen?: React.ComponentType<SvgIconProps>;
+>>>>>>> main
 }
 
 function SelectWrapper<OptionType extends OptionTypeBase, IsMulti extends boolean = false>({