+++ packages/admin/admin/src/translator/ContentTranslationServiceProvider.tsx
@@ -2,6 +2,15 @@ import * as React from "react";
 
 import { ContentTranslationServiceContext } from "./ContentTranslationServiceContext";
 
-export const ContentTranslationServiceProvider: React.FunctionComponent<ContentTranslationServiceContext> = ({ children, enabled, translate }) => {
-    return <ContentTranslationServiceContext.Provider value={{ enabled, translate }}>{children}</ContentTranslationServiceContext.Provider>;
+export const ContentTranslationServiceProvider: React.FunctionComponent<ContentTranslationServiceContext> = ({
+    children,
+    enabled,
+    showApplyTranslationDialog,
+    translate,
+}) => {
+    return (
+        <ContentTranslationServiceContext.Provider value={{ enabled, showApplyTranslationDialog, translate }}>
+            {children}
+        </ContentTranslationServiceContext.Provider>
+    );
 };