+++ packages/admin/cms-admin/src/userPermissions/hooks/currentUser.tsx
@@ -1,7 +1,10 @@
 import { gql, useQuery } from "@apollo/client";
 import { Loading } from "@comet/admin";
+<<<<<<< HEAD
 import { Typography } from "@mui/material";
 import isEqual from "lodash.isequal";
+=======
+>>>>>>> main
 import React from "react";
 
 import { ContentScopeInterface, useContentScope } from "../../contentScope/Provider";
@@ -52,7 +55,9 @@ export const CurrentUserProvider: React.FC<{
             ((user: CurrentUserInterface, permission: string, contentScope?: ContentScopeInterface) => {
                 if (user.email === undefined) return false;
                 return user.permissions.some(
-                    (p) => p.permission === permission && (!contentScope || p.contentScopes.some((cs) => isEqual(cs, contentScope))),
+                    (p) =>
+                        p.permission === permission &&
+                        (!contentScope || p.contentScopes.some((cs) => Object.entries(contentScope).every(([scope, value]) => cs[scope] === value))),
                 );
             }),
     };