+++ storybook/src/docs/components/DataGrid/DataGrid.stories.tsx
@@ -15,8 +15,13 @@ import {
     useDataGridRemote,
     usePersistentColumnState,
 } from "@comet/admin";
+<<<<<<< HEAD
 import { MoreVert } from "@mui/icons-material";
 import { Button, Menu, MenuItem, useTheme } from "@mui/material";
+=======
+import { MoreVertical } from "@comet/admin-icons";
+import { Button, Menu, MenuItem } from "@mui/material";
+>>>>>>> main
 import Box from "@mui/material/Box";
 import { DataGrid } from "@mui/x-data-grid";
 import { DataGridPro } from "@mui/x-data-grid-pro";
@@ -411,7 +416,7 @@ storiesOf("stories/components/DataGrid", module)
                     <ToolbarFillSpace />
                     <ToolbarActions>
                         <>
-                            <Button variant="text" ref={moreMenuRef} onClick={() => setShowMoreMenu(true)} endIcon={<MoreVert />} color="info">
+                            <Button variant="text" ref={moreMenuRef} onClick={() => setShowMoreMenu(true)} endIcon={<MoreVertical />} color="info">
                                 More Actions
                             </Button>
                             <Menu