+++ packages/admin/admin-rte/CHANGELOG.md
@@ -1,5 +1,6 @@
 # @comet/admin-rte
 
+<<<<<<< HEAD
 ## 7.0.0-beta.1
 
 ### Patch Changes
@@ -154,6 +155,32 @@
 -   Updated dependencies [92eae2ba9]
     -   @comet/admin@7.0.0-beta.0
     -   @comet/admin-icons@7.0.0-beta.0
+=======
+## 6.13.0
+
+### Minor Changes
+
+-   5e25348bb: Add a dialog option to the translation feature
+
+    If enabled a dialog will open when pressing the translation button showing the original text and an editable translation
+
+    Control if the dialog should be shown for the current scope via the `showApplyTranslationDialog` prop (default: true)
+
+    ```diff
+    <ContentTranslationServiceProvider
+        enabled={true}
+    +   showApplyTranslationDialog={true}
+        translate={...}
+    >
+    ```
+
+### Patch Changes
+
+-   Updated dependencies [5e25348bb]
+-   Updated dependencies [796e83206]
+    -   @comet/admin@6.13.0
+    -   @comet/admin-icons@6.13.0
+>>>>>>> main
 
 ## 6.12.0