+++ packages/admin/admin/CHANGELOG.md
@@ -1,5 +1,6 @@
 # @comet/admin
 
+<<<<<<< HEAD
 ## 7.0.0-beta.1
 
 ### Patch Changes
@@ -319,10 +320,61 @@
     >
         {/* ... */}
     </MenuItemGroup>
+=======
+## 6.13.0
+
+### Minor Changes
+
+-   5e25348bb: Add a dialog option to the translation feature
+
+    If enabled a dialog will open when pressing the translation button showing the original text and an editable translation
+
+    Control if the dialog should be shown for the current scope via the `showApplyTranslationDialog` prop (default: true)
+
+    ```diff
+    <ContentTranslationServiceProvider
+        enabled={true}
+    +   showApplyTranslationDialog={true}
+        translate={...}
+    >
+    ```
+
+-   796e83206: Add `AutocompleteField` and `AsyncAutocompleteField` components
+
+    **Examples**
+
+    ```tsx
+    <AutocompleteField
+        name="autocomplete"
+        label="Autocomplete"
+        options={[
+            { value: "chocolate", label: "Chocolate" },
+            { value: "strawberry", label: "Strawberry" },
+            { value: "vanilla", label: "Vanilla" },
+        ]}
+        getOptionLabel={(option: Option) => option.label}
+        isOptionEqualToValue={(option: Option, value: Option) => option.value === value.value}
+        fullWidth
+    />
+    ```
+
+    ```tsx
+    <AsyncAutocompleteField
+        name="asyncAutocomplete"
+        label="Async Autocomplete"
+        loadOptions={async () => {
+            // Load options here
+        }}
+        getOptionLabel={(option: Option) => option.label}
+        isOptionEqualToValue={(option: Option, value: Option) => option.value === value.value}
+        fullWidth
+    />
+>>>>>>> main
     ```
 
 ### Patch Changes
 
+<<<<<<< HEAD
 -   b5753e612: Allow partial props in the theme's `defaultProps` instead of requiring all props when setting the `defaultProps` of a component
 -   66330e4e6: Fix a bug where the `disabled` prop would not be passed to the children of `Field`
 -   Updated dependencies [803bc607f]
@@ -339,6 +391,9 @@
 -   Updated dependencies [33ba50719]
     -   @comet/admin-theme@7.0.0-beta.0
     -   @comet/admin-icons@7.0.0-beta.0
+=======
+-   @comet/admin-icons@6.13.0
+>>>>>>> main
 
 ## 6.12.0