+++ packages/admin/cms-admin/CHANGELOG.md
@@ -1,5 +1,6 @@
 # @comet/cms-admin
 
+<<<<<<< HEAD
 ## 7.0.0-beta.1
 
 ### Major Changes
@@ -34,6 +35,16 @@
 
 ### Minor Changes
 
+=======
+## 6.13.0
+
+### Minor Changes
+
+-   c51b250ca: Add loading spinner to `IFrameViewer`
+
+    This feature was added to inform users that the iframe is loading. It is particularly useful when loading takes a long time due to a slow network connection or a large amount of content or data. The feedback remains visible until the iframe is fully loaded and the `onLoad` event is triggered.
+
+>>>>>>> main
 -   dcf3f70f4: Add `overrideAcceptedMimeTypes` configuration to DAM
 
     If set, only the mimetypes specified in `overrideAcceptedMimeTypes` will be accepted.
@@ -70,6 +81,7 @@
 
 ### Patch Changes
 
+<<<<<<< HEAD
 -   @comet/admin@7.0.0-beta.1
 -   @comet/admin-date-time@7.0.0-beta.1
 -   @comet/admin-icons@7.0.0-beta.1
@@ -452,6 +464,17 @@
     -   @comet/admin-rte@7.0.0-beta.0
     -   @comet/blocks-admin@7.0.0-beta.0
     -   @comet/admin-icons@7.0.0-beta.0
+=======
+-   aee7ae4a2: Use the same logic for checking the user's content scope in Admin as it is used in the API.
+-   Updated dependencies [5e25348bb]
+-   Updated dependencies [796e83206]
+    -   @comet/admin@6.13.0
+    -   @comet/admin-rte@6.13.0
+    -   @comet/admin-date-time@6.13.0
+    -   @comet/admin-icons@6.13.0
+    -   @comet/admin-theme@6.13.0
+    -   @comet/blocks-admin@6.13.0
+>>>>>>> main
 
 ## 6.12.0